Routing reso dinamico per implementazione pagine tramite DB e maggiore personalizzazione da parte dell'utente finale
Terminate alcune visual della scheda

Bugfix minori

+++ src/libraries/Security.php
@@ -152,7 +152,7 @@ class Security
 
             #Type Int (1,2,3)
             case 'Int':
-                $data = filter_var($data, FILTER_SANITIZE_NUMBER_INT);
+                $data = (int)filter_var($data, FILTER_SANITIZE_NUMBER_INT);
                 break;
 
             #Type Float (0.1,0.2,0.3)
